+++ pypi/pgdesign/__main__.py
@@ -1,0 +1,3 @@
+from pgdesign import main
+
+main()